Allow creating project in a nonempty directory.
Currently we disallow creating a project in a directory that has any
unhidden files. This is unnecessarily restrictive, as there are only 3
files we create that could conflict. We should only disallow creation if
the directory contains one of those 3 files.

A common flow for me is to start creating models in Blender before
creating a Godot project.