Add keyboard shortcuts to lock transform axes.
While actively using a transform tool, you can press X/Y/Z to lock the
transform to an axis or (shift)+X/Y/Z to constrain the transform to a
plane. These keys are only processed if you have a transform tool
(translate/rotate/scale) active _and_ the mouse button is held.

When we handle a key during a transform, we need to accept/return so it
doesn't cascade through. In particular, without this "Y" will toggle
snap mode unintentionally.

See godotengine/godot-proposals#1215.