Add unbound shortcuts for blender-like transforms.
This adds shortcuts for "instant" transforms, where you can press the
key and immediately be transforming an object without holding the mouse
then click to commit the transformation (or ESC to abort).

This is inspired by Blender's G(rab)/R(otate)/S(cale) shortcuts, but I
decided not to add default bindings as `S` is already bound to the
regular scale tool, and it might be confusing to only bind some of them.

See godotengine/godot-proposals#1215.