Allow specifying a texture for a handle material.
This allows spatial gizmo plugins to customize the way a handle is
rendered, if desired, but does not break existing behavior (as providing
no texture uses the default).

This is useful for distinguishing different kinds of handles. For
example, the Path editor plugin can use this to distinguish curve
handles from point handles.

See godotengine/godot-proposals#1246.